core/requirements.txt
langchain
langchain-openai
beautifulsoup4
click
python-dotenv
